+++ src/components/super/i-block/providers/index.ts
@@ -12,6 +12,7 @@
  */
 
 import symbolGenerator from 'core/symbol';
+import Provider, { providers, instanceCache, requestCache, ProviderOptions } from 'core/data';
 
 import SyncPromise from 'core/promise/sync';
 import config from 'config';
@@ -26,6 +27,9 @@ import type iBlock from 'components/super/i-block/i-block';
 import type { InitLoadCb, InitLoadOptions } from 'components/super/i-block/interface';
 
 import iBlockState from 'components/super/i-block/state';
+import type { DataProviderProp } from 'components/super/i-block/providers/interface';
+
+export * from 'components/super/i-block/providers/interface';
 
 const
 	$$ = symbolGenerator();
@@ -62,6 +66,10 @@ export default abstract class iBlockProviders extends iBlockState {
 	 */
 	@hook('after:beforeDataCreate')
 	initLoad(data?: unknown | InitLoadCb, opts: InitLoadOptions = {}): CanPromise<void> {
+		if (SSR) {
+			hydrationStore.init(this.componentId);
+		}
+
 		const
 			that = this;
 
@@ -72,9 +80,10 @@ export default abstract class iBlockProviders extends iBlockState {
 		this.beforeReadyListeners = 0;
 
 		const hydrationMode =
-			HYDRATION && !this.isReadyOnce;
+			HYDRATION &&
+			hydrationStore.has(this.componentId);
 
-		if (hydrationMode && hydrationStore.has(this.componentId)) {
+		if (hydrationMode) {
 			this.state.set(hydrationStore.get(this.componentId));
 			done();
 			return;
@@ -101,12 +110,8 @@ export default abstract class iBlockProviders extends iBlockState {
 				tasks: Array<CanPromise<unknown>> = [];
 
 			if (this.state.globalName != null) {
-				const
-					storageInitialization = this.state.initFromStorage();
-
-				if (!hydrationMode) {
-					tasks.push(storageInitialization);
-				}
+				const storageInitialization = this.state.initFromStorage();
+				tasks.push(storageInitialization);
 			}
 
 			if (this.dependencies.length > 0) {
@@ -253,5 +258,79 @@ export default abstract class iBlockProviders extends iBlockState {
 		}
 
 		return Promise.resolve();
+	}
+
+	/**
+	 * Creates an instance of the DataProvider based on the specified parameters
+	 *
+	 * @param provider
+	 * @param opts
+	 */
+	createDataProviderInstance(provider: 'Provider', opts?: ProviderOptions): null;
+
+	/**
+	 * Creates an instance of the DataProvider based on the specified parameters
+	 *
+	 * @param provider
+	 * @param opts
+	 * @throws {ReferenceError} if it is not possible to create a provider based on the provided parameters
+	 */
+	createDataProviderInstance(provider: DataProviderProp, opts?: ProviderOptions): Provider;
+
+	createDataProviderInstance(provider: DataProviderProp, opts?: ProviderOptions): CanNull<Provider> {
+		const
+			that = this;
+
+		opts = {
+			...opts,
+			id: this.r.appId,
+			remoteState: this.remoteState
+		};
+
+		let
+			dp: Provider;
+
+		if (Object.isString(provider)) {
+			const
+				ProviderConstructor = <CanUndef<typeof Provider>>providers[provider];
+
+			if (ProviderConstructor == null) {
+				if (provider === 'Provider') {
+					return null;
+				}
+
+				throw new ReferenceError(`The provider "${provider}" is not defined`);
+			}
+
+			dp = new ProviderConstructor(opts);
+			registerDestructor();
+
+		} else if (Object.isFunction(provider)) {
+			const ProviderConstructor = Object.cast<typeof Provider>(provider);
+
+			dp = new ProviderConstructor(opts);
+			registerDestructor();
+
+		} else {
+			dp = <Provider>provider;
+		}
+
+		return dp;
+
+		function registerDestructor() {
+			that.r.unsafe.async.worker(() => {
+				const key = dp.getCacheKey();
+				delete instanceCache[key];
+				delete requestCache[key];
+			});
+		}
+	}
+
+	/**
+	 * Clears the component hydration data
+	 */
+	@hook('mounted')
+	protected clearComponentHydratedData(): void {
+		hydrationStore.remove(this.componentId);
 	}
 }