+++ src/components/super/i-data/i-data.ts
@@ -90,7 +90,30 @@ export default abstract class iData extends iDataHandlers {
 				this.syncDataProviderWatcher(false);
 			}
 
-			if (HYDRATION && !this.isReadyOnce && hydrationStore.has(this.componentId)) {
+			const
+				providerHydrationKey = '[[DATA_PROVIDER]]';
+
+			const setDBData = (data: CanUndef<this['DB']>) => {
+				this.saveDataToRootStore(data);
+				this.hydrationStore?.set(this.componentId, providerHydrationKey, Object.cast(data));
+				this.db = this.convertDataToDB<this['DB']>(data);
+
+				// During hydration, there may be a situation where the cache on the DB getter is set before rendering occurs,
+				// so we forcibly touch it to log this access in Vue
+				void this.db;
+			};
+
+			if (HYDRATION && hydrationStore.has(this.componentId)) {
+				const
+					store = hydrationStore.get(this.componentId),
+					data = Object.cast<CanUndef<this['DB']>>(store?.[providerHydrationKey]);
+
+				if (store != null) {
+					delete store[providerHydrationKey];
+				}
+
+				setDBData(data);
+
 				return callSuper();
 			}
 
@@ -129,7 +152,7 @@ export default abstract class iData extends iDataHandlers {
 				const db = this.convertDataToDB<this['DB']>(data);
 				void this.lfc.execCbAtTheRightTime(() => this.db = db, label);
 
-			} else if (dataProvider?.provider.baseURL != null) {
+			} else if ((!SSR || this.ssrRendering) && dataProvider?.provider.baseURL != null) {
 				const
 					needRequest = Object.isArray(dataProvider.getDefaultRequestParams('get'));
 
@@ -145,10 +168,13 @@ export default abstract class iData extends iDataHandlers {
 								return;
 							}
 
-							Object.assign(defParams[1], {
-								...label,
-								important: this.componentStatus === 'unloaded'
-							});
+							const
+								query = defParams[0],
+								opts = {
+									...defParams[1],
+									...label,
+									important: this.componentStatus === 'unloaded'
+								};
 
 							if (this.dependencies.length > 0) {
 								void this.moduleLoader.load(...this.dependencies);
@@ -158,16 +184,12 @@ export default abstract class iData extends iDataHandlers {
 								void this.state.initFromStorage();
 							}
 
-							return dataProvider.get(<RequestQuery>defParams[0], defParams[1]);
+							return dataProvider.get(<RequestQuery>query, opts);
 						})
 
 						.then(
 							(data) => {
-								void this.lfc.execCbAtTheRightTime(() => {
-									this.saveDataToRootStore(data);
-									this.db = this.convertDataToDB<this['DB']>(data);
-								}, label);
-
+								void this.lfc.execCbAtTheRightTime(() => setDBData(data), label);
 								return callSuper();
 							},