+++ src/components/super/i-static-page/modules/ss-helpers/helpers.js
@@ -7,18 +7,24 @@
  */
 
 const
-	{webpack} = require('@config/config');
+	fs = require('fs');
+
+const
+	{webpack, src, csp} = require('@config/config'),
+	getHash = include('build/hash');
+
+exports.canLoadStylesDeferred = !webpack.externalizeInline() && !csp.nonce();
 
 exports.needInline = needInline;
 
 /**
- * Returns true if should inline resources
+ * Returns true if should include the code inline in the tag
  *
  * @param {boolean} [forceInline]
  * @returns {boolean}
  */
 function needInline(forceInline) {
-	return Boolean(webpack.fatHTML() || webpack.inlineInitial() || forceInline);
+	return Boolean(forceInline || webpack.fatHTML() || webpack.inlineInitial());
 }
 
 exports.addPublicPath = addPublicPath;
@@ -62,4 +68,38 @@ function addPublicPath(path) {
 
 		return res;
 	}
+}
+
+exports.emitFile = emitFile;
+
+/**
+ * Information about output file paths
+ *
+ *  1. outputPath - absolute path to the file
+ *  2. loadPath - path to load the file, taking into account the configured public path
+ *
+ * @typedef {{
+ *   outputPath: string,
+ *   loadPath: string
+ * }} PathData
+ */
+
+/**
+ * Creates a new file with the given content and stores it in the output directory
+ *
+ * @param {string} content - file content
+ * @param {string} fileName - file name
+ *
+ * @returns {PathData}
+ */
+function emitFile(content, fileName) {
+	fileName = webpack.output({name: fileName, hash: getHash(content)});
+
+	const
+		outputPath = src.clientOutput(fileName),
+		loadPath = webpack.publicPath(fileName);
+
+	fs.writeFileSync(outputPath, content);
+
+	return {outputPath, loadPath};
 }