+++ src/core/component/init/states/before-create.ts
@@ -19,7 +19,7 @@ import { attachAccessorsFromMeta } from 'core/component/accessor';
 import { attachMethodsFromMeta, callMethodFromComponent } from 'core/component/method';
 
 import { runHook } from 'core/component/hook';
-import { globalEmitter, implementEventEmitterAPI } from 'core/component/event';
+import { implementEventEmitterAPI } from 'core/component/event';
 
 import { beforeDestroyState } from 'core/component/init/states/before-destroy';
 import { destroyedState } from 'core/component/init/states/destroyed';
@@ -167,7 +167,6 @@ export function beforeCreateState(
 	} = meta;
 
 	initFields(systemFields, component, unsafe);
-	unsafe.$async.once(globalEmitter, `destroy.${unsafe.$root.componentId}`, unsafe.$destroy.bind(unsafe));
 
 	const
 		fakeHandler = () => undefined;