+++ src/core/dom/intersection-watcher/engines/abstract.ts
@@ -109,7 +109,7 @@ export default abstract class AbstractEngine {
 		}
 
 		watcher = {
-			id: Math.random().toString().slice(2),
+			id: Object.fastHash(Math.random()),
 			target: el,
 
 			handler,