+++ src/core/prelude/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v4.?.? (2023-??-??)
+## v4.0.0-beta.19 (2023-09-08)
 
 #### :house: Internal